requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
python-dateutil==2.9.0.post0
google-api-python-client==2.140.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
tomli==2.0.1
